requirements.txt
Flask
Flask-SocketIO
google-generativeai
selenium
python-dotenv